Eliminate poetry from SQL CLI
