+++ sql-cli/noxfile.py
@@ -16,16 +16,19 @@ def dev(session: nox.Session) -> None:
     development environment to ``.nox/dev``.
     """
     session.install("nox")
-    session.install("poetry")
-    session.run("poetry", "install")
+    session.install("-e", "../python-sdk/.[all]")
+    session.install("-e", "../python-sdk/.[tests]")
+    session.install("-e", ".[all]")
+    session.install("-e", ".[tests]")
 
 
 @nox.session(python=["3.7", "3.8", "3.9"])
-@nox.parametrize("airflow", ["2.3"])
-def test(session: nox.Session, airflow) -> None:
+def test(session: nox.Session) -> None:
     """Run both unit and integration tests."""
-    session.install("poetry")
-    session.run("poetry", "install")
+    session.install("-e", "../python-sdk/.[all]")
+    session.install("-e", "../python-sdk/.[tests]")
+    session.install("-e", ".[all]")
+    session.install("-e", ".[tests]")
     # Log all the installed dependencies
     session.log("Installed Dependencies:")
     session.run("pip3", "freeze")
@@ -37,8 +40,10 @@ def test(session: nox.Session, airflow) -> None:
 @nox.session(python=["3.8"])
 def type_check(session: nox.Session) -> None:
     """Run MyPy checks."""
-    session.install("poetry")
-    session.run("poetry", "install")
+    session.install("-e", "../python-sdk/.[all]")
+    session.install("-e", "../python-sdk/.[tests]")
+    session.install("-e", ".[all]")
+    session.install("-e", ".[tests]")
     session.run("mypy", "--version")
     session.run("mypy")
