+++ sql-cli/pyproject.toml
@@ -1,14 +1,14 @@
-[tool.poetry]
+[project]
 name = "astro-sql-cli"
 version = "0.0.0"
 description = "Empower analysts to build workflows to transform data using SQL"
 authors = [
-    "Astronomer <humans@astronomer.io>",
+    {name = "Astronomer", email = "humans@astronomer.io"},
 ]
 readme = "README.md"
-license = "Apache-2.0"
+license = {file = "LICENSE"}
+requires-python = ">=3.7"
 keywords = ["sql", "transform", "analyst", "workflow", "airflow", "astronomer"]
-packages = [{include = "sql_cli"}]
 classifiers = [
     "Development Status :: 3 - Alpha",
     "License :: OSI Approved :: Apache Software License",
@@ -20,26 +20,20 @@ classifiers = [
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
 ]
-
+dependencies = [
+    "networkx",
+    "typer[all]>=0.6.1"
+]
 [project.urls]
 Home = "https://astronomer.io/"
 Source = "https://github.com/astronomer/astro-sdk/sql-cli"
 
-[tool.poetry.dependencies]
-python = ">=3.8,<4.0.0"
-typer = {extras = ["all"], version = "^0.6.1"}
-
-[tool.poetry.scripts]
-flow = "sql_cli.__main__:app"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^7.1.3"
-pytest-cov = "^4.0.0"
-mypy = "^0.982"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+[project.optional-dependencies]
+tests = [
+    "pytest",
+    "pytest-cov",
+    "mypy"
+]
 
 [tool.isort]
 src_paths = ["sql_cli", "tests"]
