+++ sql-cli/sql_cli/exceptions.py
