+++ sql-cli/sql_cli/macros/tasks.py.jinja2
