+++ sql-cli/sql_cli/sql_directory_parser.py
@@ -4,6 +4,7 @@ from pathlib import Path
 from typing import Iterable
 
 import frontmatter
+
 from sql_cli.utils import find_template_variables
