+++ sql-cli/tests/test___main__.py
@@ -1,19 +1,6 @@
-import pytest
 from typer.testing import CliRunner
 
-from sql_cli.main import generate_dag
 from sql_cli.__main__ import app
-
-
-@pytest.mark.freeze_time("2022-09-28")
-def test_generate_dag(root_directory, target_directory, dags_directory):
-    """Test that the whole DAG generation process including sql files parsing works."""
-    generate_dag(
-        directory=root_directory,
-        target_directory=target_directory,
-        dags_directory=dags_directory,
-    )
-
 
 runner = CliRunner()
